code cleanup and update dependencies
