+++ Cargo.toml
@@ -38,35 +38,35 @@ dotenvy = "0.15"
 
 # command-line interface dependencies
 clap = { version = "4.3", features = ["derive"], optional = true }
-dirs = { version = "5", optional = true }
+dirs = { version = "6", optional = true }
 humantime = { version = "2.1", optional = true }
 num_cpus = { version = "1.15", optional = true }
-tabled = { version = "0.13", optional = true }
+tabled = { version = "0.18", optional = true }
 tracing-subscriber = { version = "0.3", optional = true }
 indicatif = { version = "0.17.7", optional = true }
 futures-util = { version = "0.3.28", optional = true }
-itertools = { version = "0.12.0", optional = true }
+itertools = { version = "0.14.0", optional = true }
 tempfile = { version = "3.8", optional = true }
-which = { version = "5.0", optional = true }
+which = { version = "7.0", optional = true }
 bgpkit-commons = { version = "0.5", optional = true }
 
 # crawler dependencies
 futures = { version = "0.3", optional = true }
 oneio = { version = "0.17.0", features = ["s3"], optional = true }
 regex = { version = "1", optional = true }
-scraper = { version = "0.17", optional = true }
+scraper = { version = "0.23", optional = true }
 tokio = { version = "1", optional = true, features = ["full"] }
 
 # api dependencies
 axum = { version = "0.7", optional = true }
-tower-http = { version = "0.5", optional = true, features = ["cors"] }
+tower-http = { version = "0.6", optional = true, features = ["cors"] }
 http = { version = "1.0", optional = true }
 utoipa = { version = "4", optional = true, features = ["axum_extras", "chrono"] }
 utoipa-swagger-ui = { version = "5", optional = true, features = ["axum"] }
 axum-prometheus = { version = "0.5.0", optional = true }
 
 # database dependencies
-sqlx = { version = "0.7", features = ["runtime-tokio", "sqlite"], optional = true }
+sqlx = { version = "0.8", features = ["runtime-tokio", "sqlite"], optional = true }
 async-nats = { version = "0.34.0", optional = true }
 
 [features]
